# app/frontend/web/openapi.json
{
  "components": {
    "schemas": {
      "ContractItemSchema": {
        "description": "Schema for an item within a contract.\n\n`is_singleton` and `raw_quantity` are deliberately absent. Both are fields of ESI's\nAUTHENTICATED character/corporation contract-item routes; the public item route\nHangar Bay reads carries neither, so `is_singleton` is the mapping default and\n`raw_quantity` is NULL for every item in the corpus today. The columns stay \u2014 they\nfill in the moment a user's own contracts are ingested \u2014 but a wire field that can\nonly misinform is worse than no wire field (ESI-3).",
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "is_blueprint_copy": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Blueprint Copy"
          },
          "is_included": {
            "title": "Is Included",
            "type": "boolean"
          },
          "market_group_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Group Id"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "record_id": {
            "title": "Record Id",
            "type": "integer"
          },
          "type_id": {
            "title": "Type Id",
            "type": "integer"
          },
          "type_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type Name"
          }
        },
        "required": [
          "record_id",
          "type_id",
          "quantity",
          "is_included"
        ],
        "title": "ContractItemSchema",
        "type": "object"
      },
      "ContractListResponse": {
        "description": "A page of contracts plus the coverage figure the system_ids filter needs.\n\nA bare total hides that system_ids can only ever match the locations we resolved:\na search returning 3 results reads as \"there are 3\" when it may mean \"there are 3\nwe can place, and 40 more we cannot\". Publishing the residual next to the number\nis the convention the EVE tooling ecosystem already uses for partial data \u2014\nAdam4EVE prints a NoP (number of participants) column beside its aggregates,\nEVE Tycoon labels each row with the appraisal method behind it.",
        "properties": {
          "items": {
            "description": "List of items for the current page",
            "items": {
              "$ref": "#/components/schemas/ContractSchema"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "description": "Current page number",
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "description": "Number of items per page",
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "description": "Total number of items",
            "title": "Total",
            "type": "integer"
          },
          "unknown_system_excluded": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "How many contracts matched every other filter but were excluded because their start location has no known solar system (player-owned structures, which need an ACL-scoped token to resolve). Null when system_ids was not applied \u2014 no results were dropped for that reason, which is different from none having been.",
            "title": "Unknown System Excluded"
          }
        },
        "required": [
          "total",
          "page",
          "size",
          "items"
        ],
        "title": "ContractListResponse",
        "type": "object"
      },
      "ContractSchema": {
        "description": "Schema for a public contract.\n\n`status` and `date_completed` are deliberately absent. Both are fields of ESI's\nAUTHENTICATED character/corporation contract routes; the public route Hangar Bay\nreads carries neither, so the columns behind them hold a placeholder and a NULL for\nevery contract in the corpus today. The columns stay \u2014 they fill in the moment a\nuser's own contracts are ingested \u2014 but a wire field that can only misinform is\nworse than no wire field (ESI-3).",
        "properties": {
          "collateral": {
            "title": "Collateral",
            "type": "number"
          },
          "contract_id": {
            "title": "Contract Id",
            "type": "integer"
          },
          "date_expired": {
            "format": "date-time",
            "title": "Date Expired",
            "type": "string"
          },
          "date_issued": {
            "format": "date-time",
            "title": "Date Issued",
            "type": "string"
          },
          "end_location_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Location Id"
          },
          "for_corporation": {
            "title": "For Corporation",
            "type": "boolean"
          },
          "is_ship_contract": {
            "title": "Is Ship Contract",
            "type": "boolean"
          },
          "issuer_corporation_id": {
            "title": "Issuer Corporation Id",
            "type": "integer"
          },
          "issuer_corporation_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer Corporation Name"
          },
          "issuer_id": {
            "title": "Issuer Id",
            "type": "integer"
          },
          "issuer_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer Name"
          },
          "items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ContractItemSchema"
            },
            "title": "Items",
            "type": "array"
          },
          "price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "reward": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reward"
          },
          "start_location_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Location Id"
          },
          "start_location_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Location Name"
          },
          "start_location_system_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Location System Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "volume": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          }
        },
        "required": [
          "contract_id",
          "issuer_id",
          "issuer_corporation_id",
          "type",
          "for_corporation",
          "date_issued",
          "date_expired",
          "collateral",
          "is_ship_contract"
        ],
        "title": "ContractSchema",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SortDirection": {
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortDirection",
        "type": "string"
      },
      "SortableContractFields": {
        "description": "Fields that can be used for sorting contracts.",
        "enum": [
          "date_issued",
          "date_expired",
          "price",
          "collateral",
          "ship_name",
          "volume"
        ],
        "title": "SortableContractFields",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "API for the Hangar Bay application, providing access to EVE Online public contract data and related services.",
    "title": "Hangar Bay API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/contracts/": {
      "get": {
        "description": "Retrieves a paginated list of contracts based on specified filters.\n\nThis endpoint uses a service layer to apply advanced filtering, sorting,\nand pagination to public contracts.",
        "operationId": "list_public_contracts_contracts__get",
        "parameters": [
          {
            "description": "Case-insensitive search across contract title and ship name.",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minLength": 3,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Case-insensitive search across contract title and ship name.",
              "title": "Search"
            }
          },
          {
            "description": "Minimum price.",
            "in": "query",
            "name": "min_price",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum price.",
              "title": "Min Price"
            }
          },
          {
            "description": "Maximum price.",
            "in": "query",
            "name": "max_price",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum price.",
              "title": "Max Price"
            }
          },
          {
            "description": "Minimum collateral.",
            "in": "query",
            "name": "min_collateral",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum collateral.",
              "title": "Min Collateral"
            }
          },
          {
            "description": "Maximum collateral.",
            "in": "query",
            "name": "max_collateral",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum collateral.",
              "title": "Max Collateral"
            }
          },
          {
            "description": "Minimum runs for BPCs. (NO MATCHES \u2014 filters an always-NULL column; do not expose in clients)",
            "in": "query",
            "name": "min_runs",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": -1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum runs for BPCs. (NO MATCHES \u2014 filters an always-NULL column; do not expose in clients)",
              "title": "Min Runs"
            }
          },
          {
            "description": "Maximum runs for BPCs. (NO MATCHES \u2014 filters an always-NULL column; do not expose in clients)",
            "in": "query",
            "name": "max_runs",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": -1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum runs for BPCs. (NO MATCHES \u2014 filters an always-NULL column; do not expose in clients)",
              "title": "Max Runs"
            }
          },
          {
            "description": "Minimum Material Efficiency for BPCs. (NOT IMPLEMENTED \u2014 accepted but ignored by the service; do not expose in clients)",
            "in": "query",
            "name": "min_me",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum Material Efficiency for BPCs. (NOT IMPLEMENTED \u2014 accepted but ignored by the service; do not expose in clients)",
              "title": "Min Me"
            }
          },
          {
            "description": "Maximum Material Efficiency for BPCs. (NOT IMPLEMENTED \u2014 accepted but ignored by the service; do not expose in clients)",
            "in": "query",
            "name": "max_me",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum Material Efficiency for BPCs. (NOT IMPLEMENTED \u2014 accepted but ignored by the service; do not expose in clients)",
              "title": "Max Me"
            }
          },
          {
            "description": "Minimum Time Efficiency for BPCs. (NOT IMPLEMENTED \u2014 accepted but ignored by the service; do not expose in clients)",
            "in": "query",
            "name": "min_te",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Minimum Time Efficiency for BPCs. (NOT IMPLEMENTED \u2014 accepted but ignored by the service; do not expose in clients)",
              "title": "Min Te"
            }
          },
          {
            "description": "Maximum Time Efficiency for BPCs. (NOT IMPLEMENTED \u2014 accepted but ignored by the service; do not expose in clients)",
            "in": "query",
            "name": "max_te",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum Time Efficiency for BPCs. (NOT IMPLEMENTED \u2014 accepted but ignored by the service; do not expose in clients)",
              "title": "Max Te"
            }
          },
          {
            "description": "List of region IDs to filter by.",
            "in": "query",
            "name": "region_ids",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "List of region IDs to filter by.",
              "title": "Region Ids"
            }
          },
          {
            "description": "List of solar system IDs to filter by. Partial coverage: contracts in NPC stations carry a resolved system, contracts in player-owned structures do not and never match. The response's unknown_system_excluded field reports how many results were dropped for that reason.",
            "in": "query",
            "name": "system_ids",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "List of solar system IDs to filter by. Partial coverage: contracts in NPC stations carry a resolved system, contracts in player-owned structures do not and never match. The response's unknown_system_excluded field reports how many results were dropped for that reason.",
              "title": "System Ids"
            }
          },
          {
            "description": "List of station IDs to filter by.",
            "in": "query",
            "name": "station_ids",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "List of station IDs to filter by.",
              "title": "Station Ids"
            }
          },
          {
            "description": "List of ship type IDs to filter by.",
            "in": "query",
            "name": "type_ids",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "List of ship type IDs to filter by.",
              "title": "Type Ids"
            }
          },
          {
            "description": "Filter for contracts containing blueprints (BPCs).",
            "in": "query",
            "name": "is_bpc",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter for contracts containing blueprints (BPCs).",
              "title": "Is Bpc"
            }
          },
          {
            "description": "Filter for contracts flagged as ship contracts (contract-level flag).",
            "in": "query",
            "name": "is_ship_contract",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter for contracts flagged as ship contracts (contract-level flag).",
              "title": "Is Ship Contract"
            }
          },
          {
            "description": "Page number.",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "Page number.",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "Number of items per page.",
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "description": "Number of items per page.",
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "description": "Field to sort by.",
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortableContractFields",
              "default": "date_issued",
              "description": "Field to sort by."
            }
          },
          {
            "description": "Sort direction.",
            "in": "query",
            "name": "sort_direction",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortDirection",
              "default": "desc",
              "description": "Sort direction."
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Public Contracts",
        "tags": [
          "Contracts"
        ]
      }
    },
    "/contracts/{contract_id}": {
      "get": {
        "description": "Retrieves a single contract by its ID, including its items.",
        "operationId": "get_contract_contracts__contract_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "contract_id",
            "required": true,
            "schema": {
              "title": "Contract Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Contract",
        "tags": [
          "Contracts"
        ]
      }
    },
...
  }
}